server: Load pkcs8 private key instead of pkcs1
Let's Encrypt gives pkcs8 private keys by default. If we continue to use
pkcs1, we need to convert the private key that Let's Encrypt into a
private key that the server supports every time we get a new
certificate.

If we use pkcs8, we can just restart the service and everything will be
wonderful. Since this server is only used with Let's Encrypt right now,
this seems like the better choice.